Update convert2bvh.py
fix ValueError: cannot reshape array of size 69 into shape (24,3).
---
Using the command
```bash
emc --data config/datasets/svimage.yml --exp config/1v1p/hrnet_pare_finetune.yml --root input/ --subs input/ --ranges 0 45 1
```
generates poses that are 69-dimensional, but it is necessary to ensure that the poses array in the input JSON file is 72-dimensional. If it is 69-dimensional, you need to identify the missing three dimensions of data. These three dimensions usually correspond to the rotational information of the root joint, which might be found in the Rh field. Adjust the data accordingly to achieve a 72-dimensional poses array in your JSON file.